refactor : #263 패키지구조 수정
